+++ src/interfaces/IBondingCurve.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.8.10;
 
 interface IBondingCurve {
-    function calculateInitializationReturn(uint256 _price, uint32 _reserveRatio)
+    function calculateInitializationReturn(uint256 _price, uint256 _reserveRatio)
         external
         view
         returns (uint256);
@@ -10,14 +10,14 @@ interface IBondingCurve {
     function calculatePurchaseReturn(
         uint256 _supply,
         uint256 _poolBalance,
-        uint32 _reserveRatio,
+        uint256 _reserveRatio,
         uint256 _price
     ) external returns (uint256);
     
     function calculateSaleReturn(
         uint256 _supply,
         uint256 _poolBalance,
-        uint32 _reserveRatio,
+        uint256 _reserveRatio,
         uint256 _tokens
     ) external returns (uint256);
 }
